backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
boto3>=1.34.0
aioboto3>=12.0.0
aiosqlite>=0.19.0
gpt4all>=2.0.2
requests>=2.31.0
python-dotenv>=1.0.0
python-pptx>=0.6.21 